Improve test case
